+++ tests/browser_assert.test.js
@@ -230,4 +230,26 @@ describe("Assertions", function() {
         }, `elements failed`);
     });
 
+    it("Element", async () => {
+        await browser.open(configUrls.index);
+        await browser.assert.element("h1");
+    });
+
+    it("Element Throws", async () => {
+        await browser.open(configUrls.index);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.element("p.not-exist");
+        }, `Expected selector "p.not-exist" to find exactly 1 elements, 0 found`);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.element("p");
+        }, `Expected selector "p" to find exactly 1 elements, 2 found`);
+    });
+
+    it("Element Throws With Custom Message", async () => {
+        await browser.open(configUrls.index);
+        await utils.assertThrowsAsync(async () => {
+            await browser.assert.element("p", "element failed");
+        }, `element failed`);
+    });
+
 });